Fix fetch results inversion in search()
We are interested in the latest search() result. So far, the latest
search() results are from the lastly resolved fetch(), which is wrong
since multiple fetch() calls are not resolved in the same order.
To fix this, search() records the timestamp when called to guarantee
that the current context is still the latest call after resolving
fetch().

This implementation does *not* uses atomic operations, so the race condition bug
still (but rarely) exists. We can fix this when Atomics.compareExchange()
becomes more available.