Add an optimization for query containing many words
Word-by-word query is expensive. So, a heuristic is here: A user may
want papers containing a specific sentence, which does not need expensive
word-by-word query. If the query is long, e.g., containing more than 7
words, we infer that a user wants a specific sentence, and skip the expensive
query. This shows 30x faster results.